requirements.txt
gpytorch=1.10
networkx=3.1
numpy=1.24.3
pytorch=2.0.0
scipy=1.10.1
matplotlib=3.7.1
tqdm=4.65.0 
